requirements.txt
streamlit
google-generativeai
gradio-client
